rename view to story_translations
